same for mMax

+++ objects.hpp
@@ -438,7 +438,7 @@ struct ballGroup
 
 	inline int getMassMax()
 	{
-		int mMax = m[0];
+		double mMax = m[0];
 		for (int Ball = 0; Ball < cNumBalls; Ball++)
 		{
 			if (m[Ball] > mMax)
